Initial version that builds python package, including extension, into wheel with scikit-build-core
