+++ symmetrix/python/symmetrix/__init__.py
@@ -1,0 +1,3 @@
+__version__ = "0.0.1"
+
+from .symmetrix_calc import Symmetrix
